+++ spec/services/payola/invoice_failed_spec.rb
@@ -14,7 +14,7 @@ module Payola
 
       sub = create(:subscription, plan: plan, stripe_customer_id: customer.id, stripe_id: customer.subscriptions.first.id)
 
-      charge = Stripe::Charge.create(amount: 100, currency: 'usd', failure_message: 'Failed! OMG!')
+      charge = Stripe::Charge.create(amount: 100, currency: 'usd', failure_message: 'Failed! OMG!', customer: customer.id)
       expect(Stripe::BalanceTransaction).to receive(:retrieve).and_return(OpenStruct.new( amount: 100, fee: 3.29, currency: 'usd' ))
       event = StripeMock.mock_webhook_event('invoice.payment_failed', subscription: sub.stripe_id, charge: charge.id)